Component System WIP
Implemented part of the Component System, the main file to look at is the ECS (Entity Component System)  if you want to add more components just create a struct and give it all the parameters that'd be attached to it. Component implementation is dealt with seperate .cpp files, the ECS just organizes the components on a per entity basis so that we only have to syntactically write ECS::AddComponent<component>(entity) much like Nicks framework.